refactor(state): extract salience/outcome repositories behind the boundary (#672)
Part 2 of #672: move the raw asset_salience / asset_outcome SQL out of
src/commands/improve/{salience,outcome-loop}.ts into
src/storage/repositories/{salience,outcome}-repository.ts, mirroring the
state.db precedents (proposals-repository.ts, improve-runs-repository.ts,
events-repository.ts) — extracted verbatim, queries and row-mapping
unchanged, only relocated behind the repository boundary.

- salience-repository.ts: AssetSalienceRow, upsertAssetSalience (incl. the
  two #644 CASE guards preserving encoding_salience/encoding_source),
  getAssetSalience, getAllRankScores, recordNoOp, resetConsecutiveNoOps,
  getConsecutiveNoOps moved verbatim. isContentEncodingRow stays in
  salience.ts (pure predicate, no SQL) with the row type imported back.
- outcome-repository.ts: the upsert SQL inside updateAssetOutcome moves out
  as upsertAssetOutcome (the differential-update business logic itself
  stays in outcome-loop.ts and now calls the repository); getAssetOutcome,
  getAllAssetOutcomes, getOutcomeScoresByRef moved verbatim. The
  review_pressure omission from the conflict-update is preserved.
- New behavior: getTopRetrievalSalience(db, limit) absorbs the health read
  in commands/health/metrics.ts (previously a hardcoded LIMIT 100), with
  the existing fail-open try/catch kept at the call site.
- Both salience.ts and outcome-loop.ts re-export the moved symbols so the
  six improve importers (collapse-detector, quality-gate, preparation,
  proactive-maintenance, loop-stages, distill) and the ~11 dependent test
  files are unaffected by this move.

Verified: bunx tsc --noEmit clean; the untouched safety-net suites
(salience/salience-wiring/outcome-loop/outcome-loop-wiring/outcome-invariance,
health-ws5-observability, encoding-salience, ranking-salience-boost) all
153 tests pass with zero edits, confirming this is behavior-preserving.
