+++ src/storage/repositories/outcome-repository.ts
@@ -1,0 +1,122 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+/**
+ * Repository for the state.db `asset_outcome` table. Extracted verbatim from
+ * commands/improve/outcome-loop.ts (#672 part 2) — queries and row-mapping
+ * unchanged, only relocated behind the repository boundary. Re-exported by
+ * commands/improve/outcome-loop.ts so existing importers resolve.
+ *
+ * The differential-update business logic (warm-start seeding, EMA advance,
+ * score clipping) stays in outcome-loop.ts's `updateAssetOutcome` — only the
+ * upsert SQL that function issued moves here, as `upsertAssetOutcome`.
+ *
+ * @module outcome-repository
+ */
+
+import type { AssetOutcomeRow } from "../../commands/improve/outcome-loop";
+import type { Database } from "../database";
+
+/**
+ * Column values for one `asset_outcome` upsert. Mirrors the INSERT column
+ * list of {@link upsertAssetOutcome} verbatim.
+ */
+export interface AssetOutcomeUpsertValues {
+  ref: string;
+  lastRetrievedAt: number;
+  retrievalCount: number;
+  expectedRetrievalRate: number;
+  negativeFeedbackCount: number;
+  acceptedChangeCount: number;
+  outcomeScore: number;
+  updatedAt: number;
+}
+
+/**
+ * Upsert one `asset_outcome` row. Extracted verbatim from the tail of
+ * `updateAssetOutcome` in outcome-loop.ts — same SQL text, same bind order;
+ * the business logic that computes these values (warm-start seed vs.
+ * differential update) stays in outcome-loop.ts and calls this function.
+ *
+ * `review_pressure` is intentionally omitted from both the INSERT column list
+ * and the ON CONFLICT SET clause: the column's DEFAULT 0 seeds fresh rows, and
+ * omitting it from SET leaves an existing row's value untouched on update
+ * (never written going forward). The column itself is dropped in migration 018.
+ */
+export function upsertAssetOutcome(db: Database, values: AssetOutcomeUpsertValues): void {
+  db.prepare(
+    `INSERT INTO asset_outcome
+       (asset_ref, last_retrieved_at, retrieval_count, expected_retrieval_rate,
+        negative_feedback_count, accepted_change_count,
+        outcome_score, updated_at)
+     VALUES (?, ?, ?, ?, ?, ?, ?, ?)
+     ON CONFLICT(asset_ref) DO UPDATE SET
+       last_retrieved_at      = excluded.last_retrieved_at,
+       retrieval_count        = excluded.retrieval_count,
+       expected_retrieval_rate= excluded.expected_retrieval_rate,
+       negative_feedback_count= excluded.negative_feedback_count,
+       accepted_change_count  = excluded.accepted_change_count,
+       outcome_score          = excluded.outcome_score,
+       updated_at             = excluded.updated_at`,
+  ).run(
+    values.ref,
+    values.lastRetrievedAt,
+    values.retrievalCount,
+    values.expectedRetrievalRate,
+    values.negativeFeedbackCount,
+    values.acceptedChangeCount,
+    values.outcomeScore,
+    values.updatedAt,
+  );
+}
+
+/**
+ * Load the outcome row for one asset, or `undefined` if not yet written.
+ */
+export function getAssetOutcome(db: Database, ref: string): AssetOutcomeRow | undefined {
+  const row = db
+    .prepare(
+      `SELECT asset_ref, last_retrieved_at, retrieval_count, expected_retrieval_rate,
+              negative_feedback_count, accepted_change_count,
+              outcome_score, updated_at
+       FROM asset_outcome WHERE asset_ref = ?`,
+    )
+    .get(ref);
+  return row == null ? undefined : (row as AssetOutcomeRow);
+}
+
+/**
+ * Load ALL asset_outcome rows. Used for the proxy-adequacy tripwire computation.
+ */
+export function getAllAssetOutcomes(db: Database): AssetOutcomeRow[] {
+  return db
+    .prepare(
+      `SELECT asset_ref, last_retrieved_at, retrieval_count, expected_retrieval_rate,
+              negative_feedback_count, accepted_change_count,
+              outcome_score, updated_at
+       FROM asset_outcome ORDER BY asset_ref`,
+    )
+    .all() as AssetOutcomeRow[];
+}
+
+/**
+ * Build a Map<ref, outcome_score> for a set of refs in one query.
+ * Used by `salience.ts` to populate `outcomeSalience`.
+ */
+export function getOutcomeScoresByRef(db: Database, refs: string[]): Map<string, number> {
+  const result = new Map<string, number>();
+  if (refs.length === 0) return result;
+  const CHUNK = 500;
+  for (let i = 0; i < refs.length; i += CHUNK) {
+    const chunk = refs.slice(i, i + CHUNK);
+    const placeholders = chunk.map(() => "?").join(",");
+    const rows = db
+      .prepare(`SELECT asset_ref, outcome_score FROM asset_outcome WHERE asset_ref IN (${placeholders})`)
+      .all(...chunk) as Array<{ asset_ref: string; outcome_score: number }>;
+    for (const row of rows) {
+      result.set(row.asset_ref, row.outcome_score);
+    }
+  }
+  return result;
+}
